Refactor CharmSetSelector into more generic component
The make this component a more generic CategorySelector.
